backend/auth-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.6.0
redis==5.0.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
